2008-11-07  Hans Boehm <Hans.Boehm@hp.com> (Really Ivan Maidansky) 	(Mostly improves LLP64 support.) 	* backgraph.c, checksums.c, dbg_mlc.c, finalize.c, mark.c, 	misc.c,	reclaim.c: Changed some int and long type to word or size_t 	(and vice versa where appropriate) 	* gcj_mlc.c, include/private/dbg_mlc.h, include/private/gcconfig.h, 	include/private/thread_local_alloc.h, mark.c, 	misc.c, thread_local_alloc.c, win32_threads.c: Added intermediate 	casts to word type when casting from int to pointer (or pointer 	to int, or data pointer to code pointer) - just to remove the 	corresponding compiler warning. 	* ptr_chck.c (GC_is_visible): cast int const to word type to 	prevent left shift overflow. 	* os_dep.c: change the type of GC_mem_top_down global var 	(containing a flag) to DWORD. 	* include/gc_config_macros.h: define GC_SOLARIS_THREADS if GC_THREADS 	is defined on SunOS x86_64. 	* misc.c (GC_init_size_map): Ifdef out GC_ASSERT as a workaround 	for VC++ 2008 amd64 (v15.00.21022.08 for x64) compiler bug 	(the compiler gets hung if invoked with -Ox -D 	ALL_INTERIOR_POINTERS -D GC_ASSERTIONS) 	* backgraph.c: cast GC_gc_no value to unsigned short when 	assigned/compared to height_gc_no field of back_edges. 	* os_dep.c (GC_remove_protection): Add ARGSUSED. 	* win32_threads.c (GC_thread_exit_proc): Remove unused local var. 	* mark.c (GC_check_dirty): Move declaration out of func body.

+++ reclaim.c
@@ -341,8 +341,8 @@ STATIC int GC_n_set_marks(hdr *hhdr)
     int result = 0;
     int i;
     size_t sz = hhdr -> hb_sz;
-    int offset = MARK_BIT_OFFSET(sz);
-    int limit = FINAL_MARK_BIT(sz);
+    int offset = (int)MARK_BIT_OFFSET(sz);
+    int limit = (int)FINAL_MARK_BIT(sz);
 
     for (i = 0; i < limit; i += offset) {
         result += hhdr -> hb_marks[i];
@@ -373,7 +373,7 @@ STATIC int GC_n_set_marks(hdr *hhdr)
     int i;
     int n_mark_words;
 #   ifdef MARK_BIT_PER_OBJ
-      int n_objs = HBLK_OBJS(hhdr -> hb_sz);
+      int n_objs = (int)HBLK_OBJS(hhdr -> hb_sz);
     
       if (0 == n_objs) n_objs = 1;
       n_mark_words = divWORDSZ(n_objs + WORDSZ - 1);
